changed datetype from int to float

+++ 6/6.3.py
@@ -2,7 +2,7 @@ def convert(gallons):
     return gallons * 3.785
 
 while True:
-    gallons = int(input("Enter the number of gallons or -1 to quit"))
+    gallons = float(input("Enter the number of gallons or -1 to quit"))
     if gallons < 0:
         break
     print(f"{gallons} gallons are {convert(gallons)}")
